Bulk allocation in vmem

+++ IPv6Addressing/lib/vmem/rmem.h
@@ -8,12 +8,13 @@ Make all of your changes to main.c instead.
 #define RMEM_H
 
 #define BLOCK_SIZE 4096
+#include <stdint.h>
 
 struct rmem {
     struct sockaddr_in6 *targetIP;
     struct in6_memaddr *memList;
     int block_size;
-    int nblocks;
+    uint64_t nblocks;
 };
 
 void configure_rmem(char *filename);
